Fix CI test failures after ai-engine directory restructuring
- Fixed uvicorn app reference in main.py (src.main:app → main:app)
- Fixed import path in test_basic.py (from src.main → from main)
- Added test_imports.py to debug import issues and verify Python path setup
- Updated CI workflow to run the new import test and point to existing test files
- Fixed matrix configuration to reference correct test file paths
